requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
celery[redis]==5.4.0
redis==5.2.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
anthropic==0.43.0
streamlit==1.41.1
pandas==2.2.3
scikit-learn==1.6.1
pillow==11.1.0
rembg==2.0.63
python-dotenv==1.0.1
structlog==24.4.0
